Update README.md
In this notebook, we will be looking into a unicorn companies and perform the entire exploratory Data Analysis(EDA) process:-- from data collection to interpretation, using Python and various data analysis libraries. The goal of this EDA project is to gain insights into the company's performance, growth, and market positioning, as well as to identify any trends or patterns in the data that may be useful for future analyses. Through my analysis, I will try to uncover interesting findings that I believe will be valuable to understand about these companies how their performance.

In this notebook, you will find detailed explanations of the data cleaning and preprocessing steps, the visualizations used to explore the data, and the statistical analyses performed to validate my observations. I have also included notes on my thought process and decision-making throughout the analysis, so that anyone who views this notebook can follow along with the reasoning behind each step.

My hope is that this notebook will not only provide valuable insights into the company I have analysed but also serve as a resource for anyone looking to conduct a similar analysis in the future. Feel free to explore the notebook and reach out to me with any questions or feedback. Thank you for your time, and I hope you find this notebook informative and helpful.